2_vision/classification_segmentation/requirements.txt
streamlit_lottie==0.0.2
streamlit== 1.15.2
requests==2.28.1
Pillow==8.4.0
protobuf~=3.19.0
watchdog==2.1.8
tensorflow==2.11.0
selenium==4.7.2
leafmap==0.14.0
plotly==5.9.0
folium==0.14.0
urllib3==1.26.13